+++ src/main/java/com/dp_ua/iksparser/bot/command/BaseCommand.java
@@ -34,10 +34,10 @@ public abstract class BaseCommand implements CommandInterface {
 
     protected abstract void perform(Message message);
 
-    public int getCommandArgument(String text) {
+    public long getCommandArgument(String text) {
         if (text.startsWith("/" + command())) {
             String argument = text.substring(command().length() + 1).trim();
-            return argument.isEmpty() ? DEFAULT_NO_PAGE_ARGUMENT : Integer.parseInt(argument);
+            return argument.isEmpty() ? DEFAULT_NO_PAGE_ARGUMENT : Long.parseLong(argument);
         }
         return 0;
     }
