+++ src/main/java/com/dp_ua/iksparser/bot/command/impl/CommandCompetitionNotLoaded.java
@@ -30,7 +30,7 @@ public class CommandCompetitionNotLoaded extends BaseCommand {
     @Override
     protected void perform(Message message) {
         String chatId = message.getChatId();
-        int commandArgument = getCommandArgument(message.getMessageText());
+        long commandArgument = getCommandArgument(message.getMessageText());
         competitionFacade.showNotLoadedInfo(chatId, commandArgument, message.getEditMessageId());
     }
 }
