+++ src/main/java/com/dp_ua/iksparser/bot/command/impl/CommandSearchByName.java
@@ -30,7 +30,7 @@ public class CommandSearchByName extends BaseCommand {
     @Override
     protected void perform(Message message) {
         String chatId = message.getChatId();
-        int commandArgument = getCommandArgument(message.getMessageText());
+        long commandArgument = getCommandArgument(message.getMessageText());
         competitionFacade.startSearchByName(chatId, commandArgument, message.getEditMessageId());
     }
 }
